feat(frontend): Token 趋势展示请求数
