+++ frontend/src/features/stats/StatsCharts.test.ts
@@ -1,4 +1,4 @@
-import { sortTrendPointsByBucket, trendBucketsNeedHour } from "./StatsCharts";
+import { buildTokenTrendChartOption, sortTrendPointsByBucket, trendBucketsNeedHour } from "./StatsChartOptions";
 
 describe("StatsCharts", () => {
   it("sorts trend points by bucket time ascending", () => {
@@ -44,5 +44,33 @@ describe("StatsCharts", () => {
       { bucket: "2026-04-11T09:00:00Z", request_count: 1, input_tokens: 30, output_tokens: 3, total_tokens: 33, estimated_cost: 0.03, balance_delta: 0, quota_delta: 0 },
     ];
     expect(trendBucketsNeedHour(points)).toBe(true);
+  });
+
+  it("adds request count to token trend chart with a separate axis", () => {
+    const option = buildTokenTrendChartOption(
+      [
+        { bucket: "2026-04-11T08:00:00Z", request_count: 7, input_tokens: 20, output_tokens: 2, total_tokens: 22, estimated_cost: 0.03, balance_delta: 0, quota_delta: 0 },
+      ],
+      "zh-CN",
+      {
+        textPrimary: "#111827",
+        textSecondary: "#6a7c73",
+        border: "#e5e7eb",
+        panelSurface: "#f8fbff",
+      },
+    );
+
+    expect(option.legend).toMatchObject({ data: ["输入", "输出", "请求数"] });
+    expect(option.yAxis).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({ name: "Token" }),
+        expect.objectContaining({ name: "请求数" }),
+      ]),
+    );
+    expect(option.series).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({ name: "请求数", type: "bar", yAxisIndex: 1, data: [7] }),
+      ]),
+    );
   });
 });
